# Improve performance of generating distinct interactions