@sgratzl 2022-07-07
see also https://github.com/upsetjs/upsetjs_r/issues/14